Common: Move ColorConv to a more appropriate place.
