+++ Common/Common.vcxproj
@@ -469,12 +469,12 @@
     <ClInclude Include="Render\Text\draw_text_win.h" />
     <ClInclude Include="Serialize\SerializeDeque.h" />
     <ClInclude Include="Serialize\SerializeFuncs.h" />
-    <ClInclude Include="ColorConvNEON.h" />
+    <ClInclude Include="Data\Convert\ColorConvNEON.h" />
     <ClInclude Include="Serialize\SerializeList.h" />
     <ClInclude Include="Serialize\SerializeMap.h" />
     <ClInclude Include="Serialize\Serializer.h" />
     <ClInclude Include="CodeBlock.h" />
-    <ClInclude Include="ColorConv.h" />
+    <ClInclude Include="Data\Convert\ColorConv.h" />
     <ClInclude Include="Common.h" />
     <ClInclude Include="CommonFuncs.h" />
     <ClInclude Include="CommonTypes.h" />
@@ -705,7 +705,7 @@
     </ClCompile>
     <ClCompile Include="ArmEmitter.cpp" />
     <ClCompile Include="Buffer.cpp" />
-    <ClCompile Include="ColorConvNEON.cpp">
+    <ClCompile Include="Data\Convert\ColorConvNEON.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</ExcludedFromBuild>
@@ -791,7 +791,7 @@
     <ClCompile Include="Render\Text\draw_text_uwp.cpp" />
     <ClCompile Include="Render\Text\draw_text_win.cpp" />
     <ClCompile Include="Serialize\Serializer.cpp" />
-    <ClCompile Include="ColorConv.cpp" />
+    <ClCompile Include="Data\Convert\ColorConv.cpp" />
     <ClCompile Include="ConsoleListener.cpp" />
     <ClCompile Include="CPUDetect.cpp" />
     <ClCompile Include="MipsCPUDetect.cpp">
