+++ Common/GPU/D3D11/thin3d_d3d11.cpp
@@ -8,9 +8,9 @@
 #endif
 #include "Common/System/Display.h"
 
+#include "Common/Data/Convert/ColorConv.h"
 #include "Common/Data/Convert/SmallDataConvert.h"
 #include "Common/Data/Encoding/Utf8.h"
-#include "Common/ColorConv.h"
 #include "Common/Log.h"
 
 #include <cfloat>
