+++ GPU/D3D11/FramebufferManagerD3D11.cpp
@@ -20,11 +20,11 @@
 #include <D3Dcompiler.h>
 
 #include "Common/Common.h"
+#include "Common/Data/Convert/ColorConv.h"
 #include "Common/System/Display.h"
 #include "Common/Math/lin/matrix4x4.h"
 #include "Common/Math/math_util.h"
 #include "Common/GPU/thin3d.h"
-#include "Common/ColorConv.h"
 
 #include "Core/MemMap.h"
 #include "Core/Config.h"
