+++ GPU/GPUCommon.cpp
@@ -5,7 +5,7 @@
 
 #include "Common/Profiler/Profiler.h"
 
-#include "Common/ColorConv.h"
+#include "Common/Data/Convert/ColorConv.h"
 #include "Common/GraphicsContext.h"
 #include "Common/Serialize/Serializer.h"
 #include "Common/Serialize/SerializeFuncs.h"
