+++ GPU/Software/Rasterizer.cpp
@@ -19,10 +19,10 @@
 #include <algorithm>
 #include <cmath>
 
+#include "Common/Data/Convert/ColorConv.h"
 #include "Common/Profiler/Profiler.h"
 
 #include "Core/ThreadPools.h"
-#include "Common/ColorConv.h"
 #include "Core/Config.h"
 #include "Core/MemMap.h"
 #include "Core/Reporting.h"
