+++ android/jni/Android.mk
@@ -60,7 +60,7 @@ ARCH_FILES := \
   $(SRC)/GPU/Common/TextureDecoderNEON.cpp.neon \
   $(SRC)/Core/Util/AudioFormatNEON.cpp.neon \
   $(SRC)/Common/ArmEmitter.cpp \
-  $(SRC)/Common/ColorConvNEON.cpp.neon \
+  $(SRC)/Common/Data/Convert/ColorConvNEON.cpp.neon \
   $(SRC)/Common/Math/fast/fast_matrix_neon.S.neon \
   $(SRC)/Core/MIPS/ARM/ArmCompALU.cpp \
   $(SRC)/Core/MIPS/ARM/ArmCompBranch.cpp \
@@ -87,7 +87,7 @@ ARCH_FILES := \
   $(SRC)/GPU/Common/TextureDecoderNEON.cpp \
   $(SRC)/Core/Util/AudioFormatNEON.cpp \
   $(SRC)/Common/Arm64Emitter.cpp \
-  $(SRC)/Common/ColorConvNEON.cpp \
+  $(SRC)/Common/Data/Convert/ColorConvNEON.cpp \
   $(SRC)/Core/MIPS/ARM64/Arm64CompALU.cpp \
   $(SRC)/Core/MIPS/ARM64/Arm64CompBranch.cpp \
   $(SRC)/Core/MIPS/ARM64/Arm64CompFPU.cpp \
@@ -228,6 +228,7 @@ EXEC_AND_LIB_FILES := \
   $(SRC)/Common/Crypto/sha1.cpp \
   $(SRC)/Common/Crypto/sha256.cpp \
   $(SRC)/Common/Data/Color/RGBAUtil.cpp \
+  $(SRC)/Common/Data/Convert/ColorConv.cpp \
   $(SRC)/Common/Data/Convert/SmallDataConvert.cpp \
   $(SRC)/Common/Data/Encoding/Base64.cpp \
   $(SRC)/Common/Data/Encoding/Compression.cpp \
@@ -292,7 +293,6 @@ EXEC_AND_LIB_FILES := \
   $(SRC)/Common/UI/ViewGroup.cpp \
   $(SRC)/Common/Serialize/Serializer.cpp \
   $(SRC)/Common/ArmCPUDetect.cpp \
-  $(SRC)/Common/ColorConv.cpp \
   $(SRC)/Common/CPUDetect.cpp \
   $(SRC)/Common/ExceptionHandlerSetup.cpp \
   $(SRC)/Common/FakeCPUDetect.cpp \
